Frame class and namespace
Added a framegen namespace and fleshed out the frame class. The Frame
class is now used within the FrameGen class as data storage and
functions have been moved to more logical places. Checking a frame’s
checksums does not have to be done by a FrameGen object, for example,
although it is still possible.