Implement new Frame structure.
This update features the new 1.0 frame structure supplied by Daniel
Gastler. An unofficial COLDATA structure was used due to lack of
official information on the topic. Note: accessors and mutators of the
new structures are functional, but no automatic loading, checking,
printing or generating functions are available yet due to backward
compatibility complications.